clientHello/unmarshal: improve length checks
This commit makes a few length checks account for the two extra bytes
spent on the definitions of lengths.

It also fixes the incorrect CipherSuite length check (which was one too
many). That was inconsequential as there are usually bytes after the
CipherSuite, and the overall length was sufficient, but it's better to
be closer to what the RFC requires.