clientHello/sni: extract exactly the SNI and nothing more
RFC6066 defines the ServerList as such:
      struct {
          NameType name_type;
          select (name_type) {
              case host_name: HostName;
          } name;
      } ServerName;

      enum {
          host_name(0), (255)
      } NameType;

      opaque HostName<1..2^16-1>;

      struct {
          ServerName server_name_list<1..2^16-1>
      } ServerNameList;

This means that, in theory, more than one ServerName might fit in this
extension (although the RFC also specifies that all NameTypes shall be
distinct in a ServerNameList, and currently only defines one NameType
(value 0 as the host_name)).

The code was already trying to handle that future-proof design, but
failed to account that there might be irrelevant data _after_ the
current ServerName within the list.

This commit fixes that oversight. It is worth noting that, as we cut the
data slice at nameLen, the data slice must be checked for its length
before and not after access. In passing, we add that check to
ClientHelloBasic, where it was missing.

Strictly speaking, this does not really change anything as of the time
of this commit (for correct implementations of TLS, that is), save for
fixing the lack of length checks in the ClientHelloBasic.